kickstart fixes and tweaks
