+++ counterparty-lib/counterpartylib/lib/kickstart/utils.py
@@ -1,4 +1,5 @@
 import binascii, os, math, json, hashlib
+from multiprocessing import resource_tracker
 
 from counterparty_rs import utils as pycoin_rs_utils
 
@@ -30,3 +31,25 @@ def decode_value(key, value):
     adjusted_key = key* int(math.ceil(float(len(value))/len(key)))
     adjusted_key = adjusted_key[:len(value)]
     return bytes([_a ^ _b for _a, _b in zip(adjusted_key, value)])
+
+
+def remove_shm_from_resource_tracker():
+    """Monkey-patch multiprocessing.resource_tracker so SharedMemory won't be tracked
+
+    More details at: https://bugs.python.org/issue38119
+    """
+
+    def fix_register(name, rtype):
+        if rtype == "shared_memory":
+            return
+        return resource_tracker._resource_tracker.register(name, rtype)
+    resource_tracker.register = fix_register
+
+    def fix_unregister(name, rtype):
+        if rtype == "shared_memory":
+            return
+        return resource_tracker._resource_tracker.unregister(name, rtype)
+    resource_tracker.unregister = fix_unregister
+
+    if "shared_memory" in resource_tracker._CLEANUP_FUNCS:
+        del resource_tracker._CLEANUP_FUNCS["shared_memory"]
