+++ counterparty-lib/counterpartylib/server.py
@@ -494,12 +494,10 @@ def start_all(db):
 
 
 def reparse(db, block_index):
-    connect_to_backend()
     blocks.reparse(db, block_index=block_index)
 
 
 def rollback(db, block_index=None):
-    connect_to_backend()
     blocks.rollback(db, block_index=block_index)
